Adding ruby script to automatically check in to Rally missions
First time using Ruby and Watir. The purpose of the script is to
automatically check in to Rally missions for you so that you don't have
to spend an hour doing it manually.

As this is my first venture into using Ruby it probably doesn't follow
any common practices and thus has room for improvement.

To run the script:
$ ruby rallyAccountEmail rallyAccountPassword