Adding new version of the script
This version will open up the "boolean" mission pages in a new tab for checking in. This makes the script faster overall because it doesn't have to re-scroll through the missions page everytime it checks in those missions. Previously, because of the infinite scroller, we had to re-scroll through the missions page since we were navigating in place instead of using a new tab.