Adding solutions for HackerRank Queens Attack 2 problem in C#
The naive approach figures out the eight directions a queen can travel, and iteratively goes off in each direction, one-by-one, adding to an attack count until an obstacle or the end of the board is reached. This approach is not very fast for large values of n.

The fast approach starts out by assuming there are no obstacles, and setting the first point off the board in each direction from the queen as an obstacle. Once this is established, it goes through each obstacle and determines which direction from the queen it lies. If the obstacle is on a valid path for the queen, it will check if this obstacle is closer to the queen than any other obstacle on that path and if so, will set it as the closest obstacle. Once all closest-obstacles have been determined, it trivially calculates the distances between the queen and those obstacles. The sum of those distances is the number of attacks the queen can make.