Solution for LeftRotation problem
The solution is simple, we can do it in O(N). First create a new array for the rotation, fill it in starting at index 'd' in the array 'a' being rotated until the end of 'a'. This is because all of those numbers were shifted left by 'd' spots. The rest of the new array can be filled in from the first 'd' elements in 'a'.