Solution for Making Anagrams
Simple solution, make a dictionary mapping the characters of string 'a' to the number of times they occur in 'a'. Next, go through string 'b' and for every character not in 'a', that is a deletion. For every character already in 'a', that is one less deletion so subtract from the ocurrences in 'a's dict. Finally, look at the remaining ocurrences 'a's dict... these all willl need to be deleted.