Solution for Two Stacks problem
The goal is to figure out how to implement a queue data structure if you only have a stack data structure. The solution here uses two stacks to do this. One contains the "FIFO" order, the other contains the "LIFO" order which is needed for the queue operations. Basically, whenever we try to get the last item, which a stack doesn't have direct access to, we pop each item from the "push" stack and push them into the "pop" stack. This results in an empty push stack, and a pop stack with the elements in LIFO order. As we push new elements, they can still be pushed to the push stack, and pops will continue to pop from the pop stack until it empties again.