Solution for Contacts Trie problem
This one briefly stumped me. I took a naive approach and had an "IsContact" boolean at every node that represented the "end" of a word. This meant that when I needed to find words starting with a certain prefix, I had to brute-force through all children and their children recursively and check every node along the way to see if "IsContact" was true and then adding to a counter. This becomes very slow with a lot of contacts and with small prefix searches so it was timing out on some testcases. Then I realized... I don't actually care what words are in the structure for any Find operation, I only care about the counts. The rest was simple.