Adding Roads and Libraries solution
This one wasn't all that difficult. There is another way to solve it as well, by building out the full graph and doing a series of depth first searches. I may at some point write that implementation and see which one ends up being faster. There was one hiccup with the problem, and that was the method signature for roadsAndLibraries provided as starter code. I didn't figure it out for the longest time (really frustrating as my solution was correct) but because the parameters were all int, when trying to return "n * c_lib", it was remaining an int and wrapper around beyond MaxInt for int32. So the answer was wrong... but it would have been correct multiplying two longs.