Adding Ransom Note solution
This one is trivial, simply maintain a HashTable (dictionary in C#) of the magazine's words and their ocurrences. As long as the word was in the magazine, and the word wasn't used too many times, the magazine will work for the note.

Runtime for this would be O(M+N) where M is the number of words in the magazine and N the number of words in the note. This is because dictionary insertions and lookups are O(1).