Adding Running Median solution
This one was fun. The trick is to maintain two heaps.. a min and a max heap. The min heap will have the smallest value in the heap at the root, and the max heap will have the largest value at the root. If you keep these heaps balanced, then the median is always going to be at the root of the larger heap, or if they have the same size, the median is the sum of the two roots divided by 2.

Insertions and removals from a heap are each O(lg N) so our runtime complexity is O(N lg N) for building and maintaining our heaps. Because the median is always at the root or roots, our median lookup is constant time so our runtime complexity is O(N) for the median lookups. This means the total complexity is O(N) + O(N lg N) or O(N lg N).