Adding Running Median solution
This implementation maintains a sorted array of numbers so the median is always at the middle of the array. To keep the array sorted, when new elements are added the code searches for an insertion point using a binary search.

Because a binary search is used for insertion and median lookups are constant time, we have a runtime complexity of O(N lg N) to maintain the array's sorted order and O(N) for the median lookups. So the actual runtime complexity is O(N lg N) + O(N) or just O(N lg N).