Adding a sorted-list solution
This solution maintains a sorted list by using binary search to determine where to insert new elements. This way, the median is easily accessed in constant time and insertions are O(lg N). So, for N elements being added with the running median printed out, the runtime complexity of the program is just O(N lg N).