Adding Shortest Reach solution
This one's pretty trivial if you understand how a BFS works. The graph can be built pretty simply, just a dictionary of node numbers and a list of node numbers that are that node's direct children.

For the distances, you just need to keep track of the depth of the BFS for any given node, which is just the depth of the parent node + 1 for a child. We can use the depths from the start node to keep track of whether we have visited a note or not, so we don't end up going in circles. Since the distances are all 6, we don't need to keep track of distance; we can just multiply depth by 6 at the end.

As we are using a BFS, the runtime complexity will be O(V + E) where V is the number of vertices and E is the number of edges. Our space complexity is also O(V + E) given that we have to store the graph before calculating distances. Our distance/depth storage is only O(V).