Solution for Connected Cell problem
The code for this one is a bit extensive but the solution is relatively simple at a high level. We want to visit every cell in the grid, and if we find a cell that has a '1' value and it hasn't been visited yet (unexplored region) then we explore the region using a DFS. A BFS would also work here of course.

The runtime complexity for this would be O(N * M) since we have to visit every cell in the grid once. If all of the values in the grid were 1's, we would end up visiting every cell twice, but that is still O(N * M).