Linear solution for Ice Cream Parlor
Because of the constraints of the input for the problem. we know there is always a unique solution and the two costs corresponding to that solution will sum up to the exact amount of money. Because of this, we can simply maintain a HashTable pairing a cost and the required cost "match" to fill the difference between that cost and the money total. This allows a O(N) runtime complexity as we just run through the costs once and stop when we find a matching cost already in our HashTable.