Adding Primality solution
Another classic, testing if a number is prime. The runtime complexity for this one is O(N^(1/2)) because we only need to check divisors up to and including the floor of the square root of the number under test. As an added optimization, we skip over even numbers so it will only check half of the numbers up to the square root.